强制转换为浮点，测试
转换为浮点检测

+++ main.c
@@ -10,9 +10,9 @@ char info;
 void serialdeal() {
 	switch (info) {
 		case 'p':
-		printf("RED:%d\nGREEN:%d\nBLUE:%d\n",color_printf(1),color_printf(2),	color_printf(3));
+		printf("RED:%d\nGREEN:%d\nBLUE:%d\n",(int)color_printf(1),(int)color_printf(2),(int)color_printf(3));
 		break;
-		
+
 		case 'c':
 		color_calibrate();
 		break;
